simplify neuron_groups and afferents description
